backendML/flask_server/requirements.txt
flask==3.0.3
langchain==0.2.1
langchain_community==0.2.1
langchain_core==0.2.2rc1
geopy
python-dotenv
typing
requests==2.32.2